Battle Manager fixed succesfully
Changed the methods and variables, now it's not static anymore.
It's instantiated from scratch in the beginning of every battle and it's destroyed after the battle is finished.
Fixed the problem with the Skill and Items buttons, too.
Additionally, fixed a new discovered bug where the monsters weren't able to drop items.